chore: move feature component
